requirements.txt
google-cloud-aiplatform==1.92.0
google-cloud-storage==2.19.0
google-genai==1.14.0
gitpython==3.1.40
google-adk==0.5.0
streamlit==1.45.1
google-auth==2.29.0
requests==2.31.0
python-dotenv==1.0.0
pandas==2.2.0
